Update package UUID to match General
For some unknown reason, Julia's General package registry has listed this
package under the wrong UUID. This commit changes the UUID in Project.toml to
match the one given in the General registry.
This appears to be a mistake in General, since the registry contains the correct
tree hash which correspond to a Project.toml with a different UUID than listed
in General.
I promise I didn't reverse engineer SHA1 to pull this off.